# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Event Management API",
    "description": "API for managing events and user authentication",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "tags": [
    {
      "name": "Auth",
      "description": "User authentication and registration"
    },
    {
      "name": "Events",
      "description": "Event CRUD operations"
    }
  ],
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "description": "Creates a new user account with the provided credentials. Returns the created user info.",
        "operationId": "signup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - email already exists or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Email already exists",
                  "error": "Bad Request"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Authenticate user and get JWT token",
        "description": "Authenticates a user with email and password. Returns a JWT access token.",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 401,
                  "message": "Invalid credentials",
                  "error": "Unauthorized"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/events": {
      "get": {
        "tags": ["Events"],
        "summary": "Get all events",
        "description": "Retrieves a list of all events. No authentication required.",
        "operationId": "findAllEvents",
        "responses": {
          "200": {
            "description": "List of all events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Events"],
        "summary": "Create a new event",
        "description": "Creates a new event. Requires JWT authentication.",
        "operationId": "createEvent",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 401,
                  "message": "Unauthorized",
                  "error": "Unauthorized"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/events/{id}": {
      "get": {
        "tags": ["Events"],
        "summary": "Get a single event by ID",
        "description": "Retrieves a specific event by its ID. No authentication required.",
        "operationId": "findOneEvent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the event",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "404": {
            "description": "Event not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Event not found",
                  "error": "Not Found"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Events"],
        "summary": "Update an event",
        "description": "Partially updates an existing event by ID. Requires JWT authentication.",
        "operationId": "updateEvent",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the event",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 401,
                  "message": "Unauthorized",
                  "error": "Unauthorized"
                }
              }
            }
          },
          "404": {
            "description": "Event not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Event not found",
                  "error": "Not Found"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Events"],
        "summary": "Delete an event",
        "description": "Deletes an existing event by ID. Requires JWT authentication.",
        "operationId": "deleteEvent",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the event",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 401,
                  "message": "Unauthorized",
                  "error": "Unauthorized"
                }
              }
            }
          },
          "404": {
            "description": "Event not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Event not found",
                  "error": "Not Found"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT access token in the format: Bearer <token>"
      }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "required": ["name", "email", "password"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the user"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address (must be unique)"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "description": "Password (minimum 6 characters)"
          }
        },
        "example": {
          "name": "John Doe",
          "email": "john@example.com",
          "password": "securepass123"
        }
      },
      "LoginDto": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address"
          },
          "password": {
            "type": "string",
            "description": "Password"
          }
        },
        "example": {
          "email": "john@example.com",
          "password": "securepass123"
        }
      },
      "CreateEventDto": {
        "type": "object",
        "required": ["title", "description", "location", "date"],
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the event"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the event"
          },
          "location": {
            "type": "string",
            "description": "Location/venue of the event"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time of the event (ISO 8601 format)"
          },
          "capacity": {
            "type": "integer",
            "description": "Maximum number of attendees (defaults to 0)",
            "default": 0
          }
        },
        "example": {
          "title": "Tech Conference 2026",
          "description": "Annual technology conference featuring latest trends",
          "location": "Convention Center, Hall A",
          "date": "2026-09-15T09:00:00.000Z",
          "capacity": 500
        }
      },
      "UpdateEventDto": {
        "type": "object",
        "description": "Fields to update on an event (all optional)",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the event"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the event"
          },
          "location": {
            "type": "string",
            "description": "Location/venue of the event"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time of the event (ISO 8601 format)"
          },
          "capacity": {
            "type": "integer",
            "description": "Maximum number of attendees",
            "default": 0
          }
        },
        "example": {
          "title": "Tech Conference 2026 - Updated",
          "capacity": 750
        }
      },
      "Event": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique identifier of the event",
            "readOnly": true
          },
          "title": {
            "type": "string",
            "description": "Title of the event"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the event"
          },
          "location": {
            "type": "string",
            "description": "Location/venue of the event"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time of the event"
          },
          "capacity": {
            "type": "integer",
            "description": "Maximum number of attendees",
            "default": 0
          }
        },
        "example": {
          "id": 1,
          "title": "Tech Conference 2026",
          "description": "Annual technology conference featuring latest trends",
          "location": "Convention Center, Hall A",
          "date": "2026-09-15T09:00:00.000Z",
          "capacity": 500
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique identifier of the user",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "description": "Full name of the user"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
      "SignupResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "User created successfully"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "format": "int64"
              },
              "email": {
                "type": "string",
                "format": "email"
              }
            }
          }
        },
        "example": {
          "message": "User created successfully",
          "user": {
            "id": 1,
            "email": "john@example.com"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Login successful"
          },
          "access_token": {
            "type": "string",
            "description": "JWT access token"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "format": "int64"
              },
              "email": {
                "type": "string",
                "format": "email"
              },
              "role": {
                "$ref": "#/components/schemas/Role"
              }
            }
          }
        },
        "example": {
          "message": "Login successful",
          "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
          "user": {
            "id": 1,
            "email": "john@example.com",
            "role": "USER"
          }
        }
      },
      "Role": {
        "type": "string",
        "enum": ["USER", "ADMIN"],
        "description": "User role in the system"
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["email must be an email", "password must be longer than or equal to 6 characters"]
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        }
      }
    }
  }
}